Also prefer composition for tests
